% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epi_model_funcs.R
\name{build_SEIR}
\alias{build_SEIR}
\title{Build a simple SEIR model}
\usage{
build_SEIR(
  Sus.to.Exp = list("fitted"),
  Exp.to.Inf = list("fitted"),
  Inf.to.Rec = list("fitted"),
  detection.prob = list("fixed", 1)
)
}
\arguments{
\item{Sus.to.Exp}{A list object indicating the whether the beta transmission parameter should be \code{"fitted"} or \code{"fixed"}. If the beta transmission parameter is \code{"fixed"}, the second object in the list will be the numeric value of the parameter. If the beta transmission parameter is \code{"fitted"} no other objects are needed in the list.}

\item{Exp.to.Inf}{A list object indicating the whether the progression rate parameter should be \code{"fitted"} or \code{"fixed"}. If the progression rate parameter is \code{"fixed"}, the second object in the list will be the numeric value of the parameter. If the progression rate transmission parameter is \code{"fitted"} no other objects are needed in the list.}

\item{Inf.to.Rec}{A list object indicating the whether the recovery rate parameter should be \code{"fitted"} or \code{"fixed"}. If the recovery rate parameter is \code{"fixed"}, the second object in the list will be the numeric value of the parameter. If the recovery rate parameter is \code{"fitted"} no other objects are needed in the list.}

\item{detection.prob}{A list object indicating the whether the detection probability parameter should be \code{"fitted"} or \code{"fixed"}. If the detection probability parameter is \code{"fixed"}, the second object in the list will be the numeric value of the parameter. If the detection probability parameter is \code{"fitted"} no other objects are needed in the list. The default value for the fixed detection probability is 1.}
}
\value{
A function with the following arguments:
\itemize{
\item N = the population size.
\item pars = the initial starting values for the fitted parameters.
\item time_step = the length of the timestep in the model.
\item tau = the length of the simulation
}
}
\description{
\code{build_SEIR()} builds a SEIR model function that can be passed to \code{run_custom_model()} and \code{run_combined_model()}. The SEIR model is a simple closed structure that gives the user the ability to fix or fit the parameter values.
}
\examples{

build_SEIR(Sus.to.Exp = list("fitted"),
           Exp.to.Inf = list("fixed", 0.33),
           Inf.to.Rec = list("fixed", 0.2),
           detection.prob = list("fixed", 0.5))

}
