% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epi_model_funcs.R
\name{build_SIR}
\alias{build_SIR}
\title{Build a simple SIR model}
\usage{
build_SIR(
  Sus.to.Inf = list("fitted"),
  Inf.to.Rec = list("fitted"),
  detection.prob = list("fixed", 1)
)
}
\arguments{
\item{Sus.to.Inf}{A list object indicating the whether the beta transmission parameter should be \code{"fitted"} or \code{"fixed"}. If the beta transmission parameter is \code{"fixed"}, the second object in the list will be the numeric value of the parameter. If the beta transmission parameter is \code{"fitted"} no other objects are needed in the list.}

\item{Inf.to.Rec}{A list object indicating the whether the recovery rate parameter should be \code{"fitted"} or \code{"fixed"}. If the recovery rate parameter is \code{"fixed"}, the second object in the list will be the numeric value of the parameter. If the recovery parameter is \code{"fitted"} no other objects are needed in the list.}

\item{detection.prob}{A list object indicating the whether the detection probability parameter should be \code{"fitted"} or \code{"fixed"}. If the detection probability parameter is \code{"fixed"}, the second object in the list will be the numeric value of the parameter. If the detection probability parameter is \code{"fitted"} no other objects are needed in the list. The default value for the fixed detection probability is 1.}
}
\value{
A function with the following arguments:
\itemize{
\item N = the population size.
\item pars = the initial starting values for the fitted parameters.
\item time_step = the length of the timestep in the model.
\item tau = the length of the simulation
}
}
\description{
\code{build_SIR()} builds a SIR model function that can be passed to \code{run_custom_model()} and \code{run_combined_model()}. The SIR model is a simple closed structure that gives the user the ability to fix or fit the parameter values.
}
\examples{

build_SIR(Sus.to.Inf = list("fitted"),
          Inf.to.Rec = list("fixed", 0.2),
          detection.prob = list("fixed", 0.5))

}
