% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_final.R
\name{compute_Rs}
\alias{compute_Rs}
\title{Title}
\usage{
compute_Rs(
  covs,
  b0 = log(2),
  sd = 0.1,
  b1 = log(0.75),
  b2 = log(1.2),
  b3 = log(1.2),
  temp_scale = 4,
  temp_loc = 20,
  b4 = log(1.2)
)
}
\arguments{
\item{b4}{}
}
\description{
Title
}
