% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statmech_utils.R
\name{em_func}
\alias{em_func}
\title{a function to implement EM analysis combining statistical and mechanistic forecasting models}
\usage{
em_func(
  epi_curves,
  covdat,
  initK,
  epimdlfit,
  starting_vals,
  error_func,
  penalty_func,
  statmdlfit,
  statmdlpred,
  threshold = 20,
  max.iter = 100,
  ...
)
}
\arguments{
\item{epi_curves}{a list of epidemic curves}

\item{covdat}{a dataframe of covariate data for each geographic area}

\item{initK}{starting final size values}

\item{epimdlfit}{name of function used to fit epidemic model}

\item{starting_vals}{starting values for epidemic model}

\item{error_func}{error function to use for epi model}

\item{penalty_func}{penalty function to use for epi model}

\item{statmdlfit}{function to use to fit stat model}

\item{statmdlpred}{function to use to get predictions from stat model}

\item{threshold}{max change between iterations for convergence}

\item{max.iter}{maximum number of interations}
}
\value{
list of estimated final size K, Khist, Kstat, Kstathist,
and parameters of the epi mdl
}
\description{
a function to implement EM analysis combining statistical and mechanistic forecasting models
}
