% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_final.R
\name{em_func_model}
\alias{em_func_model}
\title{Title}
\usage{
em_func_model(
  epi_curves,
  covdat,
  pop_N,
  initK,
  epimdlfit,
  starting_vals,
  error_func,
  penalty_func,
  statmdlfit,
  statmdlpred,
  threshold = 20,
  max.iter = 100,
  cores,
  stat.family,
  tau,
  timestep = NULL
)
}
\arguments{
\item{timestep}{}
}
\description{
Title
}
