% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statmech_utils.R
\name{epi.mdl.normcurve.fit}
\alias{epi.mdl.normcurve.fit}
\title{A function to fit a normal distribution to an epidemic curve with shrinkage towards Kstat}
\usage{
epi.mdl.normcurve.fit(ecs, Kstat, prev.mdl = NULL)
}
\arguments{
\item{ecs}{the epidemic curves}

\item{Kstat}{the values of Kstat}

\item{prev.mdl}{the previous model, can be null}
}
\value{
a list with final size, peak time and spread
}
\description{
Version of the epidemiologic model that fits a normal distribution to an epidemic curve.
with shrinkage towards Kstat
(previously \code{epi.mdl.normcurve.fit.gamma})
}
