% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statmech_utils.R
\name{epiInf.EM}
\alias{epiInf.EM}
\title{Use EM to combine results from Stat and Mech models}
\usage{
epiInf.EM(
  epi.curves,
  x,
  starting.K,
  stat.mdl.fit,
  epi.mdl.fit,
  stat.mdl.pred,
  max.iter = 100,
  threshold = 10^-5
)
}
\arguments{
\item{epi.curves}{the individual epidemic curves to fit}

\item{x}{data frame of covariates. First columns assumed to contain K initialized for the odel fits}

\item{stat.mdl.fit}{a function that fits the statistical model with the following parameters and returns a fir model:
- x: a data frame of covariates that must include the final size column K a}

\item{epi.mdl.fit}{a function that calls the epi model with the following parameter and returns a fit model:
- epi.curves: the epidemic curves.
- K: the projected fornal sizes
- prev.mdl : the previous model. should behave when this is null}

\item{max.iter}{the maximum number of iterations to run}

\item{threshold}{the desired precision of the putput}

\item{stat.mdl.pred:}{a function that takes in a statisitcal model and returns a vector of predicted final sizes}

\item{epi.mdl.pred:}{a fuctnion that takes in the epdiemic model and and returns a vector of predicted final sizes}
}
\value{
a vector of estimated final sizes
}
\description{
Functoin that fits a general epidemic and statistical model using an EM
algorithm.
}
