% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statmech_utils.R
\name{epimdl.norm.init}
\alias{epimdl.norm.init}
\title{A function to fit a normal distribution to an epidemic curve}
\usage{
epimdl.norm.init(ecs)
}
\arguments{
\item{ecs}{the epidemic curves}
}
\value{
a list with final size, peak time and spread
}
\description{
Version of the epidemiologic model that fits a normal distribution to an epidemic curve.
Initialization fits all parameters, including final size, peak time (i.e., mean) and
spread (i.e., standard deviation) (formerly \code{epimdl.norm.init.beta})
}
