% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_final.R
\name{fit_epi_model}
\alias{fit_epi_model}
\title{Title}
\usage{
fit_epi_model(
  ecs,
  N,
  epi_mdl_func,
  epi_mdl_pred,
  epi_mdl_pars,
  error_func,
  priorfunc = NULL,
  prior = NULL,
  cores,
  tau,
  timestep
)
}
\arguments{
\item{timestep}{}
}
\description{
Title
}
