% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_final.R
\name{fit_norm_model}
\alias{fit_norm_model}
\title{Title}
\usage{
fit_norm_model(
  ecs,
  epi_mdl_func,
  epi_mdl_pars,
  error_func,
  priorfunc = NULL,
  prior = NULL,
  cores,
  tau,
  timestep = NULL
)
}
\arguments{
\item{timestep}{}
}
\description{
Title
}
