% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error_funcs.R
\name{gaussian_error}
\alias{gaussian_error}
\title{Gaussian error function penalizing towards 0}
\usage{
gaussian_error(ec, pred, estK)
}
\arguments{
\item{ec}{A numeric vector providing the incidence at each timestep}

\item{pred}{A numeric vector providing predicted incidence at the each timestep}

\item{estK}{A numeric object providing the estK parameter used in the mechanistic model prediction}
}
\value{
A numeric object representing the error of the objective function
}
\description{
\code{gaussian_error()} provides an error values based on two terms: i) the sum of the distance between the observed incidence of the epidemic and the predicted incidence at each timestep; and ii) the distance between the total predicted incidence and 0. This error is then used by the optimization method to fit the model to the observed data.
}
\examples{

gaussian_error(ec = observed_cases, pred = model_cases,
               estK = sum(model_cases))

}
