% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epi_model_funcs.R
\name{gaussian_mod}
\alias{gaussian_mod}
\title{Simulate a an epidemic curve using a Gaussian distribution}
\usage{
gaussian_mod(pars, times)
}
\arguments{
\item{pars}{A vector providing, in order:
\itemize{
\item the estimated epidemic size
\item the estimated peak time of the epidemic
\item the estimated spread of the epidemic
}}

\item{times}{A numeric object providing the number of timesteps that the epidemic should be simulated for}
}
\value{
A vector of incidence values for each timestep
}
\description{
\code{gaussian_mod()} creates an epidemic curve using a Gaussian distribution, given the following parameters: i) the estimated epidemic size, ii) the estimated peak time of the epidemic, and iii) estimated spread of the epidemic; as well as the length of the epidemic.
}
\examples{

gaussian_mod(pars = c(250, 15, 5), time = 30)

}
