% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assess_funcs.R
\name{get_cases_prevented}
\alias{get_cases_prevented}
\title{Calculate cases averted by allocation of a vaccine}
\usage{
get_cases_prevented(
  method.pred,
  method.name,
  dat = flat_res,
  maxvax = 3e+06,
  coverage,
  vax.eff
)
}
\arguments{
\item{method.pred}{A mathematical expression used to estimate the number of remaining cases}

\item{method.name}{A character object providing the name of the method}

\item{dat}{A dataframe providing the location name (\code{loc}), the true epidemic size (\code{K}), the observed cases (\code{observed}), the predicted sizes from the models, and the population size for the location (\code{N}).}

\item{maxvax}{A numeric object providing the max number of vaccine doses that can be allocated. The default is \code{3e6}.}

\item{coverage}{A numeric object providing the coverage of the vaccine as a proportion (i.e. between 0 and 1).}

\item{vax.eff}{A numeric object providing the vaccine efficacy as a proportion (between 0 and 1).}
}
\value{
A dataframe containing the following columns, that shows where
doses were allocated and the number of infections prevented:
- \code{loc} which provides the location name.
- \code{method} which provides the name of the method used to
calculated the estimated remaining cases.
- \code{cumv2} which provides the cumulative number of vaccine doses
allocated after each location.
- \code{inf_prevented} which provides the cumulative number of
infections/cases averted
}
\description{
\code{get_cases_prevented()} calculates the cases that could be averted if allocation of a vaccine was prioritized using the total cases expected. Specifically it uses the total number of cases expected and the current number of cases to determine which location have the most cases remaining.
}
\examples{

get_cases_prevented(method.pred = stat - observed, method.name = "stat",
                    dat = my_dataset, maxvax = 3e6, coverage = 0.75,
                    vax_eff = 0.75)

}
