% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_final.R
\name{get_infPrevented_test}
\alias{get_infPrevented_test}
\title{Title}
\usage{
get_infPrevented_test(
  method.pred,
  method.name,
  dat = flat_res,
  maxvax = 3e+06,
  loc_limit = 50000
)
}
\arguments{
\item{loc_limit}{}
}
\description{
Title
}
