% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_trueK}
\alias{get_trueK}
\title{Extract final epidemic size from the epidemic data}
\usage{
get_trueK(dat, case.col, groups)
}
\arguments{
\item{dat}{A data frame providing the epidemic cases over time for each location}

\item{case.col}{A character object providing the variable for the cases}

\item{groups}{A character object providing the variable for the location level at which the final epidemic size should be calculated}
}
\value{
A data frame providing the location and the corresponding final epidemic size
}
\description{
\code{get_trueK()} calculates the final epidemic size from epidemic data, given the column name containing the cases and the grouping variable.
}
\examples{

 get_trueK(dat = epidemic_data, case.col = n_cases,
           groups = district)

}
