% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_model_funcs.R
\name{linear_pred}
\alias{linear_pred}
\title{Predict epidemic size using a fitted glm model}
\usage{
linear_pred(mdl, x)
}
\arguments{
\item{mdl}{A glm object providing the model fit}

\item{x}{A data frame providing the covariates for each location}
}
\value{
A vector of predicted final sizes
}
\description{
\code{linear_pred()} predicts the final epidemic size for locations given a fitted glm model (generated using \code{linear_fit()}) and covariates to input into the model.
}
