% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_final.R
\name{next_state_analysis}
\alias{next_state_analysis}
\title{Title}
\usage{
next_state_analysis(
  cur_state,
  beta_i1,
  beta_i2,
  sigma,
  omega,
  phi,
  prob_detect,
  time_step,
  N
)
}
\arguments{
\item{N}{}
}
\description{
Title
}
