% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_final.R
\name{next_state_covid}
\alias{next_state_covid}
\title{Title}
\usage{
next_state_covid(
  cur_state,
  beta_p,
  beta_a,
  beta_i,
  sigma,
  phi,
  alpha_a,
  alpha_i,
  prob_detect,
  prop_asymp,
  time_step,
  N
)
}
\arguments{
\item{N}{}
}
\description{
Title
}
