% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_final.R
\name{next_state_gen}
\alias{next_state_gen}
\title{Title}
\usage{
next_state_gen(
  cur_state,
  beta_i,
  beta_a,
  sigma,
  prop_asymp,
  phi_i,
  phi_a,
  prob_detect,
  time_step,
  N
)
}
\arguments{
\item{N}{}
}
\description{
Title
}
