% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statmech_utils.R
\name{normfit_em}
\alias{normfit_em}
\title{wrapper function to fit a normal epi model within the EM algorithm}
\usage{
normfit_em(ecs, strt_vals, errorfxn, penaltyfunc, priorval)
}
\arguments{
\item{ecs}{a list of epidemic curves}

\item{strt_vals}{starting values}

\item{errorfxn}{error function}

\item{penaltyfunc}{choice of penalty function}

\item{priorval}{vector of prior values}
}
\value{
a dataframe with normal model parameters and convergence result
}
\description{
wrapper function to fit a normal epi model within the EM algorithm
}
