% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comb_model_funcs.R
\name{run_combined_model}
\alias{run_combined_model}
\title{Run the combined mechanistic-statistical EpiStat model}
\usage{
run_combined_model(
  epi.curves,
  covdat,
  pop.N,
  stat.model = "SL",
  library.SL = NULL,
  epi.model = "gaussian",
  custom.model = NULL,
  starting.vals,
  optim.method = "Nelder-Mead",
  optim.control = NULL,
  error.func,
  penalty.func,
  tau = NULL,
  timestep = NULL,
  stat.family = "gaussian",
  threshold.pct = 0.05,
  max.iter = 100,
  epi.parallel = F,
  stat.parallel = F,
  cores = NULL
)
}
\arguments{
\item{epi.curves}{A list object providing the observed or masked epidemic curves. Each element of the list object is a numeric vector.}

\item{covdat}{A data frame object providing the covariate values for each location.}

\item{pop.N}{A numeric vector providing the population size for each location.}

\item{stat.model}{A character string providing the statistical methodology that should be used. The options are a SuperLearner \code{"SL"} (the default) or a linear regression \code{"linear"}.}

\item{library.SL}{A list object providing the SL algorithms and screeners to be used in the Superlearner. Using the option \code{NULL} which will use the default set of algorithms and screeners.}

\item{epi.model}{A character string providing the epidemic model that should be used.The options are a Gaussian model \code{"gaussian"} or a custom model \code{"custom"}.}

\item{custom.model}{A function object providing the custom epidemic model if applicable. If using the Gaussian model, this argument can be left as the default \code{NULL}.}

\item{starting.vals}{A data frame providing the starting parameter values for the mechanistic model.}

\item{optim.method}{A character object providing the optimization method that should be used to fit the epi model to the data. Options are the same as those available in base R's \code{optim()}.}

\item{optim.control}{A list object providing additional parameters to control the optimization. Options are the same as those available in base R's \code{opitm()}.}

\item{error.func}{A function object providing the error calculation for the optimization of the mechanistic model.}

\item{penalty.func}{A function object providing the penalty calculation used to penalize the mechanistic model optimization towards the results from the statistical model.}

\item{tau}{A character object providing the number of timesteps that should be simulated for the mechanistic model prediction. If using the gaussian model, allow tau to be the default \code{NULL}.}

\item{timestep}{A numeric object providing the number of days in each time step. For example, a weekly time step would be \code{timestep = 7}.If using the gaussian model, allow timestep to be the default \code{NULL}.}

\item{stat.family}{A character object providing the error distribution family for the statistical model. The options are \code{"gaussian"} and \code{"poisson"}.}

\item{threshold.pct}{A numeric object providing the threshold that must be met for the model to stop. This given as a proportion (i.e. between 0 and 1) and relates to the proportion of the observed cases for each location. The value is compared to the difference between the epi and stat model components at each iteration. Once the difference falls below the proportion the model/framework will stop.}

\item{max.iter}{A numeric object providing the maximum number of iterations that should be completed. If the threshold has not been met by this number of iterations, the iterative process will stop. The default is \code{100}}

\item{epi.parallel}{A logical object indicating if the epidemic model should be run in parallel. The default is \code{FALSE}.}

\item{stat.parallel}{A logical object indicating if the statistical model should be run in parallel. The default is \code{FALSE}}

\item{cores}{A numeric object providing the number of cores that should be assigned to run the function in parallel. If not running in parallel this can be left as the default \code{NULL}.}
}
\value{
A list object containing the following objects:
\itemize{
\item \code{K} Most recent prediction of epidemic size from combined model
\item \code{Kmech} Most recent prediction of epidemic size from the mechanistic component
\item \code{Khist} Epidemic size predictions from the combined model at each iteration
\item \code{Khist.mech} Epidemic size predictions from mechanistic component at each iteration
\item \code{epi.params} Most recently optimized model parameters
\item \code{params} Optimized model parameters for the last epidemic iteration
\item \code{converged} Convergence results from the most recent iteration
\item \code{diff} Iteration difference between the last iteration and the one previous
}
}
\description{
\code{run_combined_model()} implements the EpiStat model framework - a framework combining mechanistic and statistical models to improve prediction of final epidemic size. The outputs from each model type are used to inform the next in an iterative fashion until a maximum nubmer of iterations is met or the outputs converge.
}
\examples{

run_combined_model(epi.curves = my_curves, covdat = my_data,
                   pop.N = population, stat.model = "SL",
                   library.SL = NULL, epi.model = "gaussian",
                   custom.model = NULL, starting.vals = my_vals,
                   optim.method = "Nelder-Mead",
                   optim.control = list(maxit = 1000),
                   error.func = poisson_error, penalty.func = poispen,
                   tau = 60, timestep = 7, stat.family = "gaussian",
                   threshold.pct = 0.05, max.iter = 100, epi.parallel = F,
                   stat.parallel = F, cores = NULL)


}
