% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epi_model_funcs.R
\name{run_custom_model}
\alias{run_custom_model}
\title{Fit a custom epidemic model to epidemic data}
\usage{
run_custom_model(
  ecs,
  N,
  epi.mdl.func,
  epi.mdl.pars,
  error.func,
  prior.func = NULL,
  prior = NULL,
  tau,
  timestep,
  cores = NULL,
  optim.method = "Nelder-Mead",
  optim.control = NULL
)
}
\arguments{
\item{ecs}{A list object providing the observed or masked epidemic curves. Each element of the list object is a numeric vector.}

\item{N}{A numeric vector providing the population size in each location.}

\item{epi.mdl.func}{A function object providing the function used to simulate the mechanistic model.}

\item{epi.mdl.pars}{A dataframe providing the parameter values for each parameter (columns) and location (rows).}

\item{error.func}{A function object providing the error calculation for the optimization of the mechanistic model.}

\item{prior.func}{A function object providing the penalty calculation used to penalize the mechanistic model optimization towards the results from the statistical model.}

\item{prior}{A numeric vector providing the results from the statistical model that the \code{prior.func} will penalize towards.}

\item{tau}{A character object providing the number of timesteps that should be simulated for the mechanistic model prediction.}

\item{timestep}{A numeric object providng the number of days in each time step. For example, a weekly time step would be \code{timestep = 7}.}

\item{cores}{A numeric object providing the number of cores that should be assigned to run the function in parallel. If not running in parallel this can be left as the default \code{NULL}.}

\item{optim.method}{A character object providing the optimization method that should be used to fit the epi model to the data. Options are the same as those available in base R's \code{optim()}.}

\item{optim.control}{A list object providing additional parameters to control the optimization. Options are the same as those available in base R's \code{opitm()}.}
}
\value{
A list object consisting of, for each location:
- a dataframe with the updated parameter values
- a numeric object indicating whether the optimization
converged. A \code{0} indicates convergence.
}
\description{
\code{run_custom_model()} fits a user defined custom epidemic model to epidemic data to predict final epidemic size, using general purpose optimization via \code{optim()}.
}
\examples{

run_custom_model(ecs = my_curves, epi.mdl.func = my_epi_model,
               epi.mdl.pars = starting_values, error.func = norm_error,
               prior.func = NULL, prior = NULL, tau = 70,
               timestep = 7, cores = NULL, optim.method = "Nelder-Mead",
               optim.control = list(maxit = 1000))

}
