% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epi_model_funcs.R
\name{run_gaussian_model}
\alias{run_gaussian_model}
\title{Fit a gaussian curve to epidemic data}
\usage{
run_gaussian_model(
  ecs,
  epi.mdl.pars,
  error.func,
  prior.func = NULL,
  prior = NULL,
  cores = NULL,
  optim.method = "Nelder-Mead",
  optim.control = NULL
)
}
\arguments{
\item{ecs}{A list object providing the observed or masked epidemic curves. Each element of the list object is a numeric vector.}

\item{epi.mdl.pars}{A dataframe providing the parameter values for each parameter (columns) and location (rows).}

\item{error.func}{A function object providing the error calculation for the optimization of the mechanistic model.}

\item{prior.func}{A function object providing the penalty calculation used to penalize the mechanistic model optimization towards the results from the statistical model.}

\item{prior}{A numeric vector providing the results from the statistical model that the \code{priorfunc} will penalize towards.}

\item{cores}{A numeric object providing the number of cores that should be assigned to run the function in parallel. If not running in parallel this can be left as the default \code{NULL}.}

\item{optim.method}{A character object providing the optimization method that should be used to fit the epi model to the data. Options are the same as those available in base R's \code{optim()}.}

\item{optim.control}{A list object providing additional parameters to control the optimization. Options are the same as those available in base R's \code{opitm()}.}
}
\value{
A list object consisting of, for each location:
- a dataframe with the updated parameter values
- a numeric object indicating whether the optimization
converged. A \code{0} indicates convergence.
}
\description{
\code{run_gaussian_model()} fits a gaussian curve to epidemic data to estimate the final epidemic size, using general purpose optimization via \code{optim()}.
}
\examples{

run_gaussian_model(ecs = my_curves, epi_mdl_pars = starting_values,
                   error_func = norm_error, prior_func = NULL,
                   prior = NULL)

}
