% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_final.R
\name{sim_analysis_em}
\alias{sim_analysis_em}
\title{Title}
\usage{
sim_analysis_em(
  ecs,
  pop_N,
  strt_vals,
  errorfxn,
  penaltyfunc,
  priorval,
  cores,
  tau,
  timestep
)
}
\arguments{
\item{timestep}{}
}
\description{
Title
}
