% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_final.R
\name{simulate_outbreak}
\alias{simulate_outbreak}
\title{Title}
\usage{
simulate_outbreak(
  N,
  covariates,
  pop_max = 1e+06,
  pop_min = 1000,
  init_I = 10,
  timestep = 7,
  tau
)
}
\arguments{
\item{tau}{}
}
\description{
Title
}
