% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_model_funcs.R
\name{SL_pred}
\alias{SL_pred}
\title{Predict final epidemic sizes using a trained SuperLearner}
\usage{
SL_pred(mdl, x)
}
\arguments{
\item{mdl}{A \code{SuperLearner} object providing the trained ensemble model.}

\item{x}{A dataframe providing the variables/features for each location.}
}
\value{
A numeric vector of predicted epidemic final sizes for each location.
}
\description{
\code{SL_pred()} predicts the final epidemic sizes for locations given a trained SuperLearner objected (generated by \code{SL_fit()}) and covariates to be input into the SuperLearner.
}
\examples{

SL_pred(mdl = SL_model, x = my_vars)

}
