% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error_funcs.R
\name{sqrtpen_diff}
\alias{sqrtpen_diff}
\title{Square root penalty function penalizing towards the statistical prediction}
\usage{
sqrtpen_diff(estK, prior)
}
\arguments{
\item{estK}{A numeric object providing the estK parameter used in the mechanistic model prediction}

\item{prior}{A numeric object providing the prediction from the statistical model}
}
\value{
A numeric object representing the prior penalty error
}
\description{
\code{sqrtpen_diff()} provides a penalty for the optimization based on the distance between the total predicted incidence from the epi model and the predicted incidence for the stat model.This penalty value is combined with the value produced by the error functions during the optimization.
}
\examples{

sqrtpen_diff(estK = sum(model_cases), prior = stat_cases)

}
