% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statmech_utils.R
\name{stabfunc}
\alias{stabfunc}
\title{a function to stabilize poisson error function
built in a and currently NOT USED}
\usage{
stabfunc(estK)
}
\description{
a function to stabilize poisson error function
built in a and currently NOT USED
}
