% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statmech_utils.R
\name{stat.mdl.sl.pred.ex}
\alias{stat.mdl.sl.pred.ex}
\title{A function to extrapolate from SL model prediction}
\usage{
stat.mdl.sl.pred.ex(mdl, x, tau)
}
\arguments{
\item{mdl}{the fit model}

\item{x}{a dataset with the variables in \code{mdl}}

\item{tau}{end of follow up (time)}
}
\value{
a vector of predicted final sizes
}
\description{
Function for doing the predict for doing the prediction for the
results from a stat.mdl.sl.fit (with extrapolation)
Use this function for prediction when using the stat model alone.
}
