% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_model_funcs.R
\name{stat.mdl.sl.pred}
\alias{stat.mdl.sl.pred}
\title{Predict final epidemic sizes using a trained SuperLearner}
\usage{
stat.mdl.sl.pred(mdl, x)
}
\arguments{
\item{mdl}{A \code{SuperLearner} object providing the trained ensemble model.}

\item{x}{A dataframe providing the variables/features for each location.}
}
\value{
A numeric vector of predicted epidemic final sizes for each location.
}
\description{
Predict final epidemic sizes using a trained SuperLearner
}
\examples{

stat.mdl.sl.pred(mdl = SL_model, x = my_vars)

}
