% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_final.R
\name{compute_Rs}
\alias{compute_Rs}
\title{Compute R0 values from simulated covariate data}
\usage{
compute_Rs(
  covs,
  b0 = log(2),
  sd = 0.1,
  b1 = log(0.75),
  b2 = log(1.2),
  b3 = log(1.2),
  temp_scale = 4,
  temp_loc = 20,
  b4 = log(1.2)
)
}
\arguments{
\item{covs}{Dataframe providing the covariate values for each location}

\item{b0}{Numeric object providing the R0 intercept on the log scale. The default is \code{log(2)}}

\item{sd}{Numeric object providing the standard deviation in R0 on the log scale. The default is \code{0.1}.}

\item{b1}{Numeric object providing the coefficient for the wat/san access covariate. The default is \code{log(0.75)}.}

\item{b2}{Numeric object providing the coefficient for the standing water covariate. The default is \code{log(1.2)}}

\item{b3}{Numeric object providing the coefficient for the interaction of wat/san access and standing water covariates. The default is \code{log(1.2)}}

\item{temp_scale}{Numeric object providing a scale parameter for a logistic distribution of the temperature covariate. The default is \code{4}}

\item{temp_loc}{Numeric object providing a location parameter for a logistic distribution of the temperature covariate. The default is \code{20}}

\item{b4}{Numeric object providing the coefficient for the temperature covariate. The default is \code{log(1.2)}}
}
\value{
A numeric vector providing the calculated R0 values for each location
}
\description{
Compute R0 values from simulated covariate data
}
\examples{

compute_Rs(covariate_data, b0 = log(2), sd = .1,
           b1 = log(.75), b2 = log(1.2), b3 = log(1.2),
           temp_scale = 4, temp_loc = 20, b4 = log(1.2))

}
