% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_final.R
\name{get_infPrevented_beta}
\alias{get_infPrevented_beta}
\title{Title}
\usage{
get_infPrevented_beta(method.pred, method.name, dat = flat_res, maxvax = 3e+06)
}
\arguments{
\item{maxvax}{}
}
\description{
Title
}
