% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_final.R
\name{get_trueK}
\alias{get_trueK}
\title{Extract the final epidemic size from the epidemic data}
\usage{
get_trueK(dat, groups)
}
\arguments{
\item{dat}{A data frame providing the epidemic cases over time for each location}

\item{groups}{A character object providing the variable for the location level at which the final epidemic size should be calculated}
}
\value{
A data frame providing the location and the corresponding final epidemic size
}
\description{
Extract the final epidemic size from the epidemic data
}
\examples{

 get_trueK(dat = epidemic_data, groups = "district")

}
