% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_model_funcs.R
\name{linear_fit}
\alias{linear_fit}
\title{Fit a generalized linear model}
\usage{
linear_fit(x, y, stat.family = "gaussian")
}
\arguments{
\item{x}{A data frame of covariates where each observation is a location}

\item{y}{A numeric vector object providing the outcome}
}
\value{
A glm object containing the model fit
}
\description{
\code{linear_fit()} fits a generalized linear model based on location covariates and the observed cases at each location.
}
