% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error funcs.R
\name{norm_error2}
\alias{norm_error2}
\title{Gaussian error term penalized towards the observed}
\usage{
norm_error2(ec, pred, estK)
}
\arguments{
\item{ec}{A numeric vector providing the incidence at each timestep}

\item{pred}{A numeric vector providing predicted incidence at the each timestep. All values must be positive and non-zero.}

\item{estK}{A numeric object providing the estK parameter used in the mechanistic model. All values must be positive and non-zero.}
}
\value{
A numeric object representing the error of the objective function
}
\description{
Gaussian error term penalized towards the observed
}
\examples{

norm_error2(ec = observed_cases, pred = model_cases, estK = params[1])

}
