% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statmech_utils.R
\name{normmdl}
\alias{normmdl}
\title{a function to fit a normal model}
\usage{
normmdl(par, times)
}
\arguments{
\item{par}{a vector of starting values (final size, peak time, spread)}

\item{times}{a vector of times to compute predictions}
}
\value{
a vector of predictions at \code{times}
}
\description{
a function to fit a normal model
}
