% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error funcs.R
\name{normpen}
\alias{normpen}
\title{Gaussian penalty term}
\usage{
normpen(estK, prior)
}
\arguments{
\item{estK}{A numeric object providing the estK parameter used in the mechanistic model prediction}

\item{prior}{A numeric object providing the prediction from the statistical model}
}
\value{
A numeric object representing the prior penalty error
}
\description{
Gaussian penalty term
}
\examples{

normpen(estk = pars[1], prior = kstat[1])

}
