% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statmech_utils.R
\name{normpen}
\alias{normpen}
\title{a function to make a normal penalty}
\usage{
normpen(estK, prior, sd = 1)
}
\arguments{
\item{estK}{estimated final size}

\item{prior}{prior estimate of final size}
}
\value{
log probability to be added to error function as a penalty
}
\description{
a function to make a normal penalty
}
