% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_final.R
\name{plot_mask_curves}
\alias{plot_mask_curves}
\title{Plot the masked epidemic curves}
\usage{
plot_mask_curves(dat, maskdat, X, plot_group, count, legend = FALSE)
}
\arguments{
\item{dat}{A data frame providing the full epidemic data over time for each location.}

\item{maskdat}{A data frame providing the masked epidemic data over time for each location.}

\item{X}{A character object providing the column name for the xaxis variable. This variable should be in Date format}

\item{plot_group}{A character objecting providing the column name for the variable that the data will be grouped by.}

\item{count}{A character object providing the column name for the count variable}

\item{legend}{A logical object indicating whether a legend key should be shown. The default is \code{TRUE}.}
}
\value{
A ggplot of the masked epidemic curves over the x-axis variable and grouped by the plot_group variable.
}
\description{
Plot the masked epidemic curves
}
\examples{

plot_true_curves(dat = epidemic_data, maskdat = masked_data,
                 X = "epiweek_date", plot_group = "district", count = "cases",
                 legend = TRUE)

}
