% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_final.R
\name{plot_true_curves}
\alias{plot_true_curves}
\title{Plot the full epidemic curve}
\usage{
plot_true_curves(dat, X, plot_group, count, legend = FALSE)
}
\arguments{
\item{dat}{A data frame providing the epidemic data over time for each location}

\item{X}{A character object providing the column name for the x-axis variable. This should be a variable that is in a Date format.}

\item{plot_group}{A character objecting providing the colunm name for the variable that the data will be grouped by}

\item{count}{A character object providing the column name for the count variable}

\item{legend}{A logical object indicating whether a legend key should be shown. The default is \code{TRUE}}
}
\value{
A ggplot of the epidemic curves over the x-axis variable and grouped by the plot_group variable
}
\description{
Plot the full epidemic curve
}
\examples{

 plot_true_curves(dat = epidemic_data, X = "epiweek_date",
                  plot_group = "district", count = "cases", legend = TRUE)

}
