% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_final.R
\name{poispen}
\alias{poispen}
\title{Possion penalty term}
\usage{
poispen(estK, prior)
}
\arguments{
\item{estK}{A numeric object providing the estK parameter used in the mechanistic model prediction}

\item{prior}{A numeric object providing the prediction from the statistical model}
}
\value{
A numeric object representing the prior penalty error
}
\description{
Possion penalty term
}
\examples{

poispen(estK = params[1], prior = stat_result)

}
