% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statmech_utils.R
\name{poispen}
\alias{poispen}
\title{a function to make a poisson penalty}
\usage{
poispen(estK, prior)
}
\arguments{
\item{estK}{estimated final size}

\item{prior}{prior estimate of final size}
}
\value{
log probability to be added to error function as a penalty
}
\description{
a function to make a poisson penalty
}
