% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statmech_utils.R
\name{poisson_error}
\alias{poisson_error}
\title{A function to compute a poisson error function}
\usage{
poisson_error(ec, pred, estK)
}
\arguments{
\item{ec}{a vector of case counts}

\item{pred}{a vector of predictions}

\item{estK}{estimated final size}
}
\value{
log probability under poisson error structure
}
\description{
A function to compute a poisson error function
}
