% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error_funcs.R
\name{poisson_error2}
\alias{poisson_error2}
\title{Poisson error function penalizing towards observed cases}
\usage{
poisson_error2(ec, pred, estK)
}
\arguments{
\item{ec}{A numeric vector providing the incidence at each timestep}

\item{pred}{A numeric vector providing predicted incidence at the each timestep. All values must be positive and non-zero.}

\item{estK}{A numeric object providing the estK parameter used in the mechanistic model. All values must be positive and non-zero.}
}
\value{
A numeric object representing the error of the objective function
}
\description{
#' @description
\code{poisson_error2()} provides an error values based on two terms: i) the sum of the distance between the observed incidence of the epidemic and the predicted incidence at each timestep; and ii) the distance between the total predicted incidence and the total observed cases. This error is then used by the optimization method to fit the model to the observed data.
}
\examples{

poisson_error2(ec = observed_cases, pred = model_cases, estK = sum(model_cases))

}
