% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_final.R
\name{sim_covariates}
\alias{sim_covariates}
\title{Title}
\usage{
sim_covariates(
  N,
  poverty = "average",
  climate = c(0.33, 0.33, 0.33),
  temperature = c(30, 20, 35),
  water_san = "average",
  stand_water = "average",
  education = "average"
)
}
\arguments{
\item{education}{}
}
\description{
Title
}
