% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_model_funcs.R
\name{SL_fit}
\alias{SL_fit}
\title{Train a SuperLearner ensemble}
\usage{
SL_fit(
  x,
  y,
  family = "gaussian",
  CV_folds = 10,
  library.SL = NULL,
  parallel = F,
  cores = NULL
)
}
\arguments{
\item{x}{A dataframe providing the variables for each location.}

\item{y}{A numeric vector providing the outcome for each location.}

\item{family}{A character object providing the error distribution. Currenly only \code{gaussian} (the default) and \code{binomial} are supported.}

\item{library.SL}{A list object providing the SL algorithms and screeners to be used in the Superlearner. Using the option \code{NULL} which will use the default set of algorithms and screeners.}

\item{parallel}{A logical object indicating whether the SuperLearner should be run in parallel. The default is \code{FALSE}.}

\item{cores}{A numeric object providing the number of cores to be used in the parallel process. The default value is \code{NULL}.}
}
\value{
A \code{SuperLearner} object containing the trained model
}
\description{
\code{SL_fit()} trains a SuperLearner machine learning ensemble on outbreak data. There are options to define the library of algorithms and functionality to run in parallel using multiple cores.
}
\examples{

SL_fit(x = my_vars, y = epi_size, family = "gaussian",
           library.SL = my_library, parallel = F, cores = NULL)

}
