% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statmech_utils.R
\name{stat.mdl.lin.fit}
\alias{stat.mdl.lin.fit}
\title{Function for fitting a simple linear statistical model appropriate for
passing in to em_func}
\usage{
stat.mdl.lin.fit(x, y)
}
\arguments{
\item{x}{the data frame of covariates to fit based on}

\item{y}{outcome}
}
\value{
a fit model
}
\description{
Function for fitting a simple linear statistical model appropriate for
passing in to em_func
}
