% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_final.R
\name{stat.mdl.lin.pred}
\alias{stat.mdl.lin.pred}
\title{Function for doing the predict for doing the predict for the
results from stat.mdl.lin.fit}
\usage{
stat.mdl.lin.pred(mdl, x)
}
\arguments{
\item{mdl}{the fit model}

\item{x}{covariate matrix}
}
\value{
a vector of predicted final sizes
}
\description{
Function for doing the predict for doing the predict for the
results from stat.mdl.lin.fit
}
