% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_final.R
\name{stat.mdl.sl.fit.para2}
\alias{stat.mdl.sl.fit.para2}
\title{Train a SuperLearner ensemble in parallel}
\usage{
stat.mdl.sl.fit.para2(x, y, cores = 1, family = "gaussian")
}
\arguments{
\item{x}{A dataframe providing the variables for each location.}

\item{y}{A numeric vector providing the outcome for each location.}

\item{cores}{A numeric object providing the number of cores to be used in the parallel process. The defaultt value is \code{1}.}

\item{family}{A character object providing the error distribution. Currenly only \code{gaussian} (the default) and \code{poisson} are supported.}
}
\value{
A \code{SuperLearner} object containing the trained model
}
\description{
Train a SuperLearner ensemble in parallel
}
