% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_final.R
\name{stat.mdl.sl.fit}
\alias{stat.mdl.sl.fit}
\title{Title}
\usage{
stat.mdl.sl.fit(x, y, family = "gaussian")
}
\arguments{
\item{x}{A dataframe providing the variables for each location.}

\item{y}{A numeric vector providing the outcome for each location.}

\item{family}{A character object providing the error distribution. Currenly only \code{gaussian} (the default) and \code{poisson} are supported.}
}
\value{
A \code{SuperLearner} object containing the trained model
}
\description{
Title
}
\examples{

stat.mdl.sl.fit(x = my_vars, y = epi_size, family = "gaussian")

}
