% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statmech_utils.R
\name{stat.mdl.sl.fit}
\alias{stat.mdl.sl.fit}
\title{A function to fit SuperLearner to observed epidemic sizes}
\usage{
stat.mdl.sl.fit(x, y)
}
\arguments{
\item{x}{the data frame of covariates to fit based on and outcome K}
}
\value{
a fit superlearner model
}
\description{
Function for doing superlearner fit of the K data appropriate for
passing in to epiInf.EM
previously \code{stat.mdl.sl.fit.beta}
}
