% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statmech_utils.R
\name{stat.mdl.sl.fit}
\alias{stat.mdl.sl.fit}
\title{Function for doing superlearner fit of the K data appropriate for
passing in to em_func}
\usage{
stat.mdl.sl.fit(x, y)
}
\arguments{
\item{x}{the data frame of covariates to fit based on}

\item{y}{outcome}
}
\value{
a fit superlearner model
}
\description{
Function for doing superlearner fit of the K data appropriate for
passing in to em_func
}
