% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statmech_utils.R
\name{stat.mdl.sl.pred.ex}
\alias{stat.mdl.sl.pred.ex}
\title{Function for doing the predict for doing the predict for the
results from a stat.mdl.sl.fit.beta (with extrapolation)}
\usage{
stat.mdl.sl.pred.ex(mdl, x, tau)
}
\arguments{
\item{mdl}{the fit model}
}
\value{
a vector of predicted final sizes
}
\description{
Function for doing the predict for doing the predict for the
results from a stat.mdl.sl.fit.beta (with extrapolation)
}
