% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statmech_utils.R
\name{stat.mdl.sl.pred}
\alias{stat.mdl.sl.pred}
\title{Predict final sizes using SuperLearner model}
\usage{
stat.mdl.sl.pred(mdl, x)
}
\arguments{
\item{mdl}{the fit model}

\item{x}{dataframe with all variables in \code{mdl}}
}
\value{
a vector of predicted final sizes
}
\description{
Function for predicting final epidemic sizes using results from
stat.mdl.sl.fit
}
