% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statmech_utils.R
\name{stat.mdl.sl.pred}
\alias{stat.mdl.sl.pred}
\title{Function for doing the predict for doing the predict for the
results from a stat.mdl.sl.fit}
\usage{
stat.mdl.sl.pred(mdl, x)
}
\arguments{
\item{mdl}{the fit model}
}
\value{
a vector of predicted final sizes
}
\description{
Function for doing the predict for doing the predict for the
results from a stat.mdl.sl.fit
}
